Remove Core.java
Removes Core.java, migrating it's fuctionality into GlobalGameData.java
All references to Core.java have been replaced by GlobalGameData.java

Also resizes test galaxy to 10X10, so there are only 100 sectors,
instead of 2500.
