+++ app/src/main/java/dadeindustries/game/gc/model/Faction.java
@@ -1,4 +1,5 @@
 package dadeindustries.game.gc.model;
 
 public enum Faction {
+	UNITED_PLANETS, INIATS, GHZRGORZ, MORPHERS
 }
